Basic packaging set up
This sets up the basic packaging structure for Safir. It has a PEP 518
structure with a pyproject.toml. Most metadata is in the setup.cfg file.
We use setuptools_scm to set version info.

Note that importlib_metadata is a conditional import because it's part
of the standard library in Python 3.8.