Tox-based testing and linting
This tox set up is heavily based on
https://hynek.me/articles/testing-packaging/ and
https://github.com/python-attrs/attrs/blob/master/tox.ini

The benefit of tox over running pytest directly is that in principle
even multi-python testing can be done locally (especially if I adopt
pyenv). tox runs each test in its own virtual environment.

In this set up, we run pytest, wrapped in coverage. The coverage-report
tox environment gathers the test coverage from each environment so that,
for example, code branches that depend on the version of Python get
reported properly. The specific configuration of coverage is based on
the hynek.me article.

This set up removes the reliances on a lot of pytest extensions. mypy
now runs in its own environment separate from pytest. Linting (flake8)
now happens in a special linting environment. For that, we run
pre-commit, which fails if any of the precommit tasks fail (even in
black has to modify a file!).