Get tox parallel mode to work
Run tox environments in parallel with "tox -p all"

All that needed to be done was ensure that coverage-report would happen
after the individual coverage test runs. This is done with the "depends"
configuration. As we test more versions of Python, we'll just have to
update that listing.

For more info see
https://tox.readthedocs.io/en/latest/example/basic.html#parallel-mode