Cache .tox in github actions
Caching the .tox directory should speed up tests. The strategy here is
that the .tox cache is keyed by the hash of the .pyproject.toml,
setup.cfg, and .pre-commit-config.yaml files. If any of these change,
then the project dependencies or testing set up are likely to have
changed, and thus it's work rebuilding the .tox environment from
scratch.