Add logging middleware
The bind_logger middleware binds a logger to the handler with context
about the request, such as a UUID for the request, along with the path
and HTTP method. This lets us collect log messages for the handling of a
specific endpoint, or even specific request.

This is based on the logging set up developed for sqrbot-jr.

The tests use the https://github.com/aio-libs/pytest-aiohttp pytest
plugin to let us easily construct clients for aiohttp.web applications
that we create in tests.