Fix duplicated stdout when using safir.logging
Each time safir was configured, it stacked another handler into
the safir logger.  This meant that if logging was repeatedly
configured in the same Python process, all log output to stdout
would be duplicated.  This primarily affected test suites, where
each test would reinitialize logging.  The impact was that failed
tests from files with lots of tests would have log out duplicated
a number of times equal to the number of tests, which made it hard
to find the error or understand the output.

Avoid this by clearing the handlers in the safir logger each time
logging is configured to ensure that only the single, wanted handler
is added.