Fix typing on Python 3.7
When using the backport of importlib.metadata, the return type of
metadata() confuses mypy since it implements __getitem__ and mypy
isn't smart enough to know that means it supports dict operations.
Declare it as email.message.Message (the correct type in Python 3.8
and later) and alias that to Any for Python 3.7.