Optionally add a timestamp to log messages
Kubernetes tracks the timestamp of each output line internally, so
this is not strictly necessary for our Kubernetes services, but
Argo CD doesn't display that timestamp and other log aggregation may
not necessarily.  Add an optional argument to configure_logging to
allow an application to request a timestamp be added to each log
message.