Remove aiohttp support
Remove most of the aiohttp server support.  Change the documentation
to say that Safir is for FastAPI applications instead.

This does not yet add any FastAPI support, just removes the most
obvious aiohttp bits.  It does not update the documentation for how
to use some aiohttp middleware.  That will be done separately when
they are replaced with FastAPI dependencies.