Delay AsyncClient creation and allow recreation
The approach of creating one AsyncClient on startup and destroying
that one client without a way of recreating it on shutdown works
fine for the application when run normally, but fails in test
scenarios that reuse the application.

Instead, defer creation of the AsyncClient until the first time
it's requested, clear it on aclose(), and recreate it if it's asked
for again.  This allows applications to aclose() the client on
shutdown but still work when the same app object is reused for
multiple tests.