Update GitHub Actions workflow
Don't try to upload documentation on pull requests unless the
pull request is from a ticket branch.  This avoids failed
documentation uploads from dependabot PRs.

Update the CI workflow to match the current recommendation, so that
the workflow is run on pull requests, pre-commit is run with its
own action, and we cache the tox environments.