Put log level in severity JSON key
Google Log Explorer automatically understands the severity key in
JSON logging, but structlog's standard processor puts the level
into the level key.  Add a new processor used only for JSON logging
(to not break the formatting of the console logger) that uses the
structlog processor to get the correct log level but adds it as the
severity key instead.