Add Kubernetes support code
Add a safir.kuberentes module with an initialize_kubernetes method
that contains the correct logic for choosing an initialization
method.

Start a new safir.testing.* namespace for test support code, and
add safir.testing.kubernetes with support for mocking the
kubernetes_asyncio API.  This currently imports the API stubs from
Gafaelfawr and moneypenny with some generalization.

Tell isort to add trailing commas so that it doesn't fight with
black.