Improve handling of X-Forwarded-Proto
In the XForwardedMiddleware, rather than storing the forwarded
scheme from X-Forwarded-Proto in the request state, simply update
the scope of the request so that Starlette will think the incoming
scheme was the sheme from X-Forwarded-Proto.  Drop the state
variable that was previously set.

This will ensure that other Starlette functions such as url_for
will pick up the correct scheme and thus generate URLs that
correctly use https even if the service is behind an ingress that
terminates TLS.