Update documentation based on code review
Move the FastAPI db_session dependency above the documentation for
create_async_session, since it is the normal pattern.  Make that
clear with a note in the create_async_session section.  Remove all
mention of running tests directly with pytest and recommend using
tox -e py with an explicit list of tests.  Avoid code in the middle
of sentences in the DatabaseSessionDependency class documentation.