Pass the engine into initialize_database
Rather than having initialize_database always create a new engine
and return it, making the caller responsible for disposing of it,
have the caller create the engine and pass it into
initialize_database.  This allows test suites to use a single
engine for multiple tests, which improves connection pooling and
thus test performance, and also makes the interface a bit less
awkward.