Remove support for managed transactions
Managed transactions fail with FastAPI 0.74 and later, presumably
due to the yield dependency refactor, and I have been unable to
find either a minimal reproducer so that I can open a bug or find
what Gafaelfawr is doing differently.  Any attempt to call commit
either explicitly or implicitly in the database session dependency
fails by immediately raising asyncio.CancelledError.

This is probably a bug somewhere that should be fixed, but since I
can't narrow it down well enough to report it and since it's easy
enough to explicitly manage transactions (and arguably a bit
cleaner), and vo-cutouts was already doing so anyway, just remove
the support for managed transactions and always return a regular
session.  Update the documentation accordingly.