Enable setup-python action caching
Normally we don't use the cache in setup-python because it's more
effective to cache the entire tox environment.  But in the case of
the periodic validation job where we're intentionally not using
the tox cache to ensure that we get the latest versions of all
dependencies, the pip cache will be useful.

Populate the pip cache via the regular CI runs as well, so that
they benefit from the periodic CI job and vice versa even though
most of the caching will be done via the tox environment.