Add arq dependency for FastAPI
arq is a asyncio distributed Redis queue package. This adds a FastAPI
dependency for sending jobs to that queue from a FastAPI app.

This also includes a mock queue, which is useful for testing.

This code is ported from lsst-sqre/noteburst where is it currently being
used.

Note there's something wrong with our Sphinx set up capturing attribute
docstrings for dataclasses. While I'm trying to figure this out, in the
meantime this way of duplicating the attribute doc information in the
class docstring works for the time being.