Support pool_pre_ping for create_database_engine
Support passing pool_pre_ping down to the underlying
create_async_engine API call.  This tells SQLAlchemy to test the
connection whenever a new session is created.

SLAC is seeing a problem where connections are reset when used
during Gafaelfawr login to create a new token while querying the
database to determine if the user is an admin (which is the first
database query that would be done during the login process).  This
seems to be due to a connection timeout, and setting this parameter
should work around the problem.

Setting this unconditionally would be simpler, but it means a
database query every time a session is created, which for some
applications would significantly increase the database traffic and
thus have potential performance issues.  Make it optional so that
the application can choose (and possibly make it configurable in
the application as well).