Support sending Uvicorn logs through structlog
Add a new function, configure_uvicorn_logging, which overrides the
Uvicorn logging configuration to use structlog and adds processors
to add the log severity and (for access logs) the details of the
request as structlog context.  The context is formatted so that, if
logged as JSON, the fields match the expectation of Google's Cloud
Logging system.

Unfortunately, there doesn't appear to be an easy way to test this
function inside Safir's test suite, since Uvicorn is not friendly
to being run as a fixture or in the course of a test function.