Add Profile and LogLevel enums
Add enums with the acceptable values of Profile and LogLevel for
configure_logging and configure_uvicorn_logging.  Modify those
functions to accept either strings or enums, for backward
compatibility.

This change is in preparation for converting the FastAPI template
to use pydantic.BaseSettings instead of a dataclass for its
Configuration class.  These enums will be used to validate the
profile and log_level settings.