Add CamelCaseModel helper class
Add new helper class safir.pydantic.CamelCaseModel, which is
identical to pydantic.BaseModel except with configuration added
to accept camel-case keys using the safir.pydantic.to_camel_case
alias generator.  This is a bit less self-documenting, but is less
tedious if there are a lot of classes that need this configuration
(such as a complex Kubernetes custom object model).