Add a Slack webhook client and error reporting
Add a Slack client and associated models that allow posting
structured messages to Slack using the same formatting we currently
use for mobu. The models handle truncation and validation of the
messages according to the Slack API limitations.

Add a SlackException class that knows how to convert itself to a
structured Slack message for rich error reporting. This can be used
as a base class for application exceptions that track additional
application-specific context data about the source of an exception.

Add a custom FastAPI route class that reports all uncaught exceptions
to Slack, along with a SlackIgnoredException base class that excludes
exceptions from this reporting. If this class is not configured, it
silently does nothing and still re-raises the exception, so
configuring a Slack webhook is not required.

Add support for mocking a Slack webhook, taken from Gafaelfawr.

The Slack exception reporting prompted two minor additions to the
Safir datetime library: a new format_datetime_for_logging function
taken from Gafaelfawr that formats a datetime object in a more
human-friendly format than isodatetime, and the optional ability
for current_datetime to return a datetime with microseconds.