Add a prefix to reported exceptions in Slack
We plan on using a common webhook to report exceptions from multiple
apps, so we need to add the application name to the beginning of the
message.

This raises the possibility that the message will need to be
retruncated, since we're making it longer. Enable validation on
attribute assignment for the SlackMessage models to make this easy.
This is not perfect, since truncation is destructive (ideally we
would truncate on output instead), but it should do for now.