Add control over whether Slack message is verbatim
The main part of the Slack message defaulted to not verbatim, which
meant mentions and channel links were automatically expanded.  Keep
that default but add a way to control it and set it to false for
exception reports, where the text of the message comes from the
exception and may contain arbitrary characters that shouldn't be
parsed.