Redo module layout and classes for Slack webhooks
To leave space for other types of Slack support to be added in the
future, rename SlackClient to SlackWebhookClient and move it, the
route handler, and SlackIgnoredException to safir.slack.webhook.
Move the Block Kit models and the SlackException class (which is a
Block Kit aware exception) to safir.slack.blockkit. Make similar
changes to the layout of the test suite, and rename the documentation
page to be specific to webhooks.

Similarly, change MockSlack to MockSlackWebhook, since it only
handles webhooks, but don't change the safir.testing.slack module
structure since it can probably share a module with other mocks.

Create separate SlackTextField and SlackCodeField classes instead
of having two optional attributes, one of which must be set, in
SlackField, and do the same for attachments. Add a bit of class
hierarchy and an additional validator to try to keep everything
straight in SlackMessage (which due to reuse of behavior between
fields and attachments is a bit awkward).

Mention that Slack has declared attachments as legacy (despite not
offering any replacement for the critical functionality we want).

Add overloaded prototypes for format_datetime_for_logging to deal
with some typing issues. Add the Pydantic mypy plugin to resolve
some typing complaints with the new Block Kit model structure.

Improve reporting of uncaught exceptions that inherit from
SlackException.