Support blocks in Slack Block Kit messages
mobu used additional non-attachment blocks for some purposes. Add
support for that to SlackMessage so that attachments can be avoided
when the attachment semantics are not needed.

Restructure the type system for Block Kit components to have base
SlackTextBlock and SlackCodeBlock classes and make fields a
specialization of them. This allows removing some duplicate code and
Pydantic validation.

Update the documentation for some changes made during the development
of this feature. The main message is now verbatim by default, add a
warning about attachments, and mention adding respx as an application
dev dependency.