Enable more type checking for Pydantic models
This forces some irritating type annotations for Slack message
fields because of invariance of mutable sequences, but all of the
other approaches seem to be worse. It caught one error in how Slack
field maximum lengths were set that made the setting unstable
depending on how multiple inheritance was interpreted.