Exclude X-Auth-Request-User from the schema
auth_dependency and auth_logger_dependency pull the username from
the X-Auth-Request-User header, set by Gafaelfawr. By default, any
header parameters like this are also added to the OpenAPI schema of
routes that use them. However, in this case that addition is
incorrect; X-Auth-Request-User is an internal implementation detail
of Gafaelfawr-protected applications, not a header that the user
can or should supply. Suppress it from the schema accordingly.