Change validate_exactly_one_of to a root validator
Given what it does, this should have always been a root validator.
Making it an individual field validator requires being too clever,
and requires a very careful calling convention. Convert it to a
root validator while we're making backward-incompatible changes.

Make the test suite use a real Pydantic model, which is a better
test of how the validator works than calling it directly.