Use new tox v4 syntax
tox v4 added subcommands as the preferred command-line syntax.
Executing it without subcommands invokes a default subcommand
called "legacy" and the upgrading documentation recommends always
using subcommands for tox run and when there may be a conflict
with an environment name. Update the documentation to always use
the new syntax.