Add watch support for list_namespaced_pod
Allow list_namespaced_pod to be run under a watch, mostly to watch
for changes to the pod phase. Add patch_namespaced_pod_status to
change the phase, primarily for tests (application code is unlikely
to call this since pod phase is handled by Kubernetes itself).

If the resource version is not specified to a watch, the watch now
starts with the next event, not the first recorded event, to match
the behavior of the Kubernetes API.