Extract more data from httpx exceptions
Previously, SlackWebException assumed that if an httpx exception
inherited from RequestError, it would have a request attribute, and
if it didn't, it wouldn't. This turns out to be incorrect; some
RequestError exceptions may not have a request for various reasons,
and some that do not inherit from RequestError do.

Adjust the exception parsing to take this into account by attempting
to extract the request and falling back on the non-request handling
on any error.

Fix one place where SlackTextField was used instead of SlackTextBlock
for something added to the message blocks. (The different length
restriction is unlikely to matter in practice, but this is more
correct.)