Switch to scriv for change log management
scriv maintains change logs in individual fragments that can be
merged with the corresponding PR without creating merge conflicts,
and then assembles those fragments into a new change log entry
during the release process. Move the pending changes to a scriv
fragment, add a scriv configuration, and update the development and
release documentation to document use of scriv.

Also note in the preamble of the change log that Safir follows
semver.