Patch create_app_client to preset JWT
This fixes GitHubAppClientFactory.create_app_client. I mistakenly though
that the oauth_token attribute on the GitHubAPI constructor could be
used to cache the app's JWT. It turns out that the JWT is treated
entirely separately, and currently can't be cached in GitHubAPI.
Instead, the JWT needs to be passed to each request method. This commit
achieves the desired behaviour by patching the Gidgethub GitHubAPI
methods to always set the jwt keyword argument.