Add a helper function for Click command help
For the Phalanx command-line tool, I added support for a second
level of commands to the standard help boilerplate. The result is
complex enough that it seems worth lifting into Safir proper rather
than existing only in a template. The result is a new safir.click
module providing a display_help function.

Add documentation of the run_with_asyncio decorator to the new
user guide page about Click support, since this is the most common
place where that decorator is used.