Improve error reporting from docs-linkcheck
Sphinx doesn't summarize the linkcheck errors at the end of the
output, but it does write them to output.txt. Move the link checking
to a makefile target so that we can use shell trickery to output the
contents of that file on link checking failures.

Add help and clean targets to the Makefile, and remove docs/api when
building docs to avoid stale files left-over from automodsumm.