Tag safir.asyncio docs with :inherited-members:
sphinx_automodsumm's ability to discover the methods of a class
appears to have some bug with some but not all classes that inherit
from Generic that seems to be related to the use of __slots__ to
determine the methods of the class. Tagging safir.asyncio with
:inherited-members: changes the code to use dir() to retrieve all
members instead, which avoids the problem.