Add end and finished to AsyncMultiQueue
In order to support the way another version of this data structure
was used in the Nublado REST spawner, add a way to mark the queue
contents as complete without clearing it. This supports the use case
of starting a reader after the queue is complete and therefore should
signal iterators to end, allowing it to see all of the captured data
and then immediately exit without further waiting.