Fix serialization of objects in watch events
In the Kubernetes mock, objects serialized into watch events were
not being serialized properly. They were keeping the Python
snake-case attribute names instead of the Kubernetes camel-case
names. This prevented deserialization to Kubernetes model objects
and was also confusing.

Fix this by passing the serialize parameter to to_dict when
serializing the objects, and restructure the code a bit to reduce
duplicate work. Add an explicit test for proper object serialization
and deserialization.