Make Pod creation from a Job more correct in mock
Randomize the Pod name in a similar way that Kubernetes does when
the default generateName behavior is used. If an explicit name was
set in the pod template in the Job, use that name; don't replace it
with a generated name. Honor generateName set in the pod template
in the Job.