Convert to Pydantic v2
Update all Pydantic code and examples to Pydantic v2. Use
validation_alias instead of env for configuring the environment
variables in examples that use BaseSettings. Convert
validate_exactly_one_of to a model validator and require that it
be called with mode="after".

Pydantic now distinguishes between str and int, and the GitHub IDs
all appear to be ints, so change the GitHub App and webhook models
to type those fields as int rather than str.