Add watch support for namespaces
Add watch, field selector, and label selector support to
list_namespace in the Kubernetes mock.

read_namespace and list_namespace in the Kubernetes mock now
only return namespace objects that have been explicitly created,
not implicit namespaces created by creating another object without
making a namespace first. This more closely matches the behavior
of Kubernetes while still making it easy to use the mock in a test
environment simulating a pre-existing namespace.

Fix the documentation for label_selector in list_* functions to
reflect that it is applied to the regular list return, not just the
watch.