Treat GitHub app and installation IDs as ints
Gidgethub treats app and installation IDs as str, but the GitHub
API appears to return int. Use int in GitHubAppClientFactory for
consistency with what GitHub returns, and convert them to str when
passing them to Gidgethub. Add a changelog entry and a note on the
class about this out of an abundance of caution.