Fix spawn_uvicorn documentation
There was a remaining bug in one of the examples where the fixture
was yielding a non-existent variable, and the name of the fixture
conflicted with the name of a local variable. Fix both of those, and
add a try/finally block to make sure that the process is always
shut down on any errors.