Convert to FastAPI lifespan functions
As of FastAPI 0.93.0, FastAPI supports a lifespan async context
manager to do startup and shutdown. The part before the yield is
run during application startup and the part after the yield is run
during application shutdown. The older on_event annotations are now
deprecated.

Add a development dependency on FastAPI 0.93.0 and change the test
suite uses of on_event annotations to lifespan functions. Update all
of the documentation to use the lifespan function instead of on_event
handlers.

This change does not bump the Safir FastAPI dependency because Safir's
own code still works fine with either method.