Fix default_queue_name in RedisArqQueue
The initialize method of RedisArqQueue didn't pass the
default_queue_name argument to __init__, so the default queue name
was forced to the default. It was set correctly in the pool, but
the enqueue method overrides the queue name from the pool, so that
setting had no effect. This caused messages without an explicit
_queue_name argument to always be sent to the default queue regardless
of the arguments to initialize.