Support pickling of SlackException
Exception classes that call the `BaseException.__init__` method
(possibly via `Exception`) with a different number of arguments than
the `__init__` method of the derived exception class takes cannot be
pickled and unpickled. Since arq uses pickle to record the exception
of failed jobs, and we may want such exceptions to be derived from
`SlackException` so that we can easily report them to Slack, this is
an annoying limitation. The restriction is due to the properties of
the default `__reduce__` method on `BaseException` and probably cannot
be changed in Python.

Work around this by using the pattern recommended at the end of the
discussion in https://github.com/python/cpython/issues/44791, namely
avoid calling `BaseException.__init__` entirely. This means we can
no longer rely on the default `__str__` behavior and must implement
`__str__`.

Add a test that `SlackException` and classes derived from it with
different numbers of constructor arguments can be pickled and unpickled
without loss of the information that goes into `to_slack`.