Separate safir-arq from safir
Move the safir.arq module into a separate package that can be
independently distributed on PyPI. This allows the arq support
library to be installed in contexts where the Safir dependency is
too heavy-weight.

Convert all of the submodules of Safir into directories so that
they can include py.typed files now that the safir package is a
namespace-only package. When importing across the namespace package,
avoid using relative imports just for clarity. Use relative imports
only within the same subpackage of Safir.