Import the UWS support layer from vo-cutouts
Import the UWS implementation from vo-cutouts. Provide the support
code for the backend worker as safir.arq.uws, since it needs to be
installable separately and dependeded on safir.arq anyway. Provide
the rest as safir.uws, hiding as many of the implementation details
as possible (which requires importing internal modules in the test
suite, since the tests need to poke at the internals).

The UWS code depends on vo-models, which currently still uses
Pydantic v1 features, so the deprecation warning for the legacy
class configuration syntax has to be suppressed for now until
vo-models is fixed.

Add a new uws dependency group that should be used when using the
uws code. This duplicates some dependencies from arq and db, since
that code is used internally.