Allow overriding UWS arq queue dependency early
Previously, the UWS dependency had to be initialized first before
the arq queue could be overridden by the test suite. This means that
Redis has to be running, since the application will enable UWS first
with a Redis queue (which immediately tries to connect) and only
later will switch to a mock queue.

Modify the UWS dependency so that the arq queue can be overridden
before initialize and the overridden queue will be used instead,
which should allow UWS applications to be tested without a local
Redis.