Move WorkerSettings into safir.arq
The WorkerSettings dataclass in safir.arq.uws is generally useful
for any service using arq, not just UWS services. Lift it into
safir.arq and update the documentation to mention it.

Also update the documentation to reflect the fact that the settings
can be either a class or an object, and that the arq command-line
tool wants the fully-qualified name of the WorkerSettings object or
class, not just the namespace that contains it. (Both might work,
but the arq documentation now says to do the former.)