Remove exported TypeVars used for Generic
A TypeVar used as the qualifier for Generic in a class definition
is not resolved correctly by Sphinx when it comes from a private
submodule, even if that TypeVar is imported and re-exported by the
documented module. Remove the exports for the two TypeVars in that
category and instead ignore them in Sphinx.