Split safir.logging into a separate PyPI package
In preparation for adding logging configuration to UWS arq workers,
which should not depend on the full `safir` package, split
`safir.logging` into a new PyPI package `safir-logging`. `safir`
depends on this package, so this should not affect backward
compatibility.

This requires some additional complexity to the integration between
logging configuration and the logger dependency, which is done via
a conditional import and skipped if the `safir.dependencies` module
is not available.

Mention in the arq documentation that `safir.arq` is also available
via the `safir-arq` package, noticed when making a similar update for
`safir.logging`.