Add support functions for Alembic
Add functions that support use of Alembic in Safir-based
applications. These include utility functions for the Alembic
`env.py` file that work with Safir's database support, functions
to mark a database as current for Alembic when initializing it,
and a function to check whether the database schema is up-to-date
that can be used during application startup.

Add `configure_alembic_logging` to safir.logging, which configures
Alembic to use structlog and JSON for logging, aiding parsing of
the logs by Google's Cloud Logging system.

Add optional support for Alembic to the UWS library. This is not
yet documented in the manual; that will be done separately as part
of integrating support into the Safir FastAPI app template.