Reorganize UWS database schema
Using a directory named `_schema` with private modules inside it
does not work correctly with Sphinx and automodapi. Sphinx is unable
to find the imported and re-exported symbols from those private
modules.

The separation of the schema into separate files is unnecessary, so
merge them all together into a single `_schema.py` module file, which
does work correctly with Sphinx.

Convert to the new style of explicit inheritance from DeclarativeBase
and use `Mapped`, `mapped_column`, and type derivation from type
annotations instead of `Column`. This is the new, recommended way of
doing ORM definitions in SQLAlchemy. Bump the version dependency for
`safir[db]` to SQLAlchemy 2.0.0 or later since some of the elements of
this new style were only added in the runup to the 2.0.0 release.