Fix documentation build errors
The documentation pages for the Pydantic types are now generated but
not included in the table of contents, so they have to be exxluded
for now in order to avoid generating orphaned pages. Also update the
nitpick exclusions to remove some that are now unnecessary and add
another case where autodoc_pydantic generates a spurious link.