Refactor metrics configuration
Rework the metrics configuration so that `MetricsConfiguration` is
a union type supporting either a disabled or a Kafka-based metrics
configuration, which can be used as a drop-in type in an application
`BaseSettings` class and will automatically select one or the other
based on what configuration is available. Have those two configuration
classes in turn create two different `EventManager` classes that
provide the same interface, and two different underlying `EventPublisher`
classes. Use this to implement the no-op metrics functionality rather
than checking a boolean. This allows all of the Kafka configuration
to be omitted if metrics are disabled.

Update the documentation accordingly, and fix a few minor problems
with the documentation discovered as part of the Gafaelfawr integration.

Move the `app_name` field to the top level of the metrics configuration
on the grounds that this will likely apply to any application metrics,
including things that are not events. Rename `metrics_events` to just
`events`, since the metrics part is clear in context.

Use the `topic_prefix` name uniformly instead of `base_topic_prefix`.